requirements.txt
python-telegram-bot
openai
sqlalchemy
alembic
requests
configparser
python-dotenv